[Docker] Better multi-stage builds
For ll-app, build stage are less coupled.
Some repetition is introduced but it is better suited
for docker-compose automated builds.

Makefile is modified accordingly.
Dockerfile is added, and template removed,
as it doesn't need to be configured with make
anymore.

`docker-compose.yml.template` contained a wrong
build target for worker.

Now `make configure` with the registry URL
as DOCKER_ID variable should work to export images
to a local registry. (Probably ?)